backend/marketing/serializers.py
from rest_framework import serializers
from .models import Campaign


class CampaignSerializer(serializers.ModelSerializer):
    class Meta:
        model = Campaign
        fields = ["id", "name", "budget", "spend", "status", "created_at", "updated_at"]
