backend/requirements.txt
Django==5.0.6
gunicorn==22.0.0
psycopg[binary]==3.1.19
dj-database-url==2.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
pydantic
djangorestframework
pytest
pytest-django